+++ src/_utils.ts
@@ -1,17 +1,7 @@
-export function normalizeKey (key: string | undefined): string {
-  if (!key) { return '' }
-  return key.replace(/[/\\]/g, ':').replace(/^:|:$/g, '')
-}
-
-export function normalizeBase (base: string | undefined): string {
-  base = normalizeKey(base)
-  return base ? (base + ':') : ''
-}
-
 type Awaited<T> = T extends Promise<infer U> ? Awaited<U> : T;
 type Promisified<T> = Promise<Awaited<T>>
 
-export function wrapToPromise<T> (val: T): Promisified<T> {
+export function wrapToPromise<T> (val: T) {
   if (!val || typeof (val as any).then !== 'function') {
     return Promise.resolve(val) as Promisified<T>
   }