+++ src/types.ts
@@ -2,22 +2,26 @@ export type StorageValue = null | string | String | number | Number | boolean | 
 export type WatchEvent = 'update' | 'remove'
 export type WatchCallback = (event: WatchEvent, key: string) => any
 
+type MaybePromise<T> = T | Promise<T>
+
+export type Unwatch = () => MaybePromise<void>
+
 export interface StorageMeta {
-  atime?: Date,
+  atime?: Date
   mtime?: Date
   [key: string]: StorageValue | Date | undefined
 }
 
 export interface Driver {
-  hasItem: (key: string) => boolean | Promise<boolean>
+  hasItem: (key: string) => MaybePromise<boolean>
   getItem: (key: string) => StorageValue
-  setItem?: (key: string, value: string) => void | Promise<void>
-  removeItem?: (key: string) => void | Promise<void>
-  getMeta?: (key: string) => StorageMeta | Promise<StorageMeta>
-  getKeys: (base?: string) => string[] | Promise<string[]>
-  clear?: () => void | Promise<void>
-  dispose?: () => void | Promise<void>
-  watch?: (callback: WatchCallback) => void | Promise<void>
+  setItem?: (key: string, value: string) => MaybePromise<void>
+  removeItem?: (key: string) => MaybePromise<void>
+  getMeta?: (key: string) => MaybePromise<StorageMeta>
+  getKeys: (base?: string) => MaybePromise<string[]>
+  clear?: () => MaybePromise<void>
+  dispose?: () => MaybePromise<void>
+  watch?: (callback: WatchCallback) => MaybePromise<Unwatch>
 }
 
 export interface Storage {
@@ -27,7 +31,7 @@ export interface Storage {
   setItem: (key: string, value: StorageValue) => Promise<void>
   removeItem: (key: string, removeMeta?: boolean) => Promise<void>
   // Meta
-  getMeta: (key: string, nativeMetaOnly?: true) => StorageMeta | Promise<StorageMeta>
+  getMeta: (key: string, nativeMetaOnly?: true) => MaybePromise<StorageMeta>
   setMeta: (key: string, value: StorageMeta) => Promise<void>
   removeMeta: (key: string) => Promise<void>
   // Keys
@@ -35,7 +39,8 @@ export interface Storage {
   // Utils
   clear: (base?: string) => Promise<void>
   dispose: () => Promise<void>
-  watch: (callback: WatchCallback) => Promise<void>
+  watch: (callback: WatchCallback) => Promise<Unwatch>
+  unwatch: () => Promise<void>
   // Mount
   mount: (base: string, driver: Driver) => Storage
   unmount: (base: string, dispose?: boolean) => Promise<void>