+++ test/drivers/fs.test.ts
@@ -1,3 +1,4 @@
+import { describe, it, expect, vi } from 'vitest'
 import { resolve } from 'path'
 import { readFile, writeFile } from '../../src/drivers/utils/node-fs'
 import { testDriver } from './utils'
@@ -19,11 +20,29 @@ describe('drivers: fs', () => {
         expect(meta.size).toBeGreaterThan(0)
       })
       it('watch filesystem', async () => {
-        const watcher = jest.fn()
+        const watcher = vi.fn()
         await ctx.storage.watch(watcher)
         await writeFile(resolve(dir, 's1/random_file'), 'random')
         await new Promise(resolve => setTimeout(resolve, 500))
         expect(watcher).toHaveBeenCalledWith('update', 's1:random_file')
+      })
+
+      const invalidKeys = [
+        '../foobar',
+        '..:foobar',
+        '../',
+        '..:',
+        '..'
+      ]
+      for (const key of invalidKeys) {
+        it('disallow path travesal: ' , async () => {
+          await expect(ctx.storage.getItem(key)).rejects.toThrow('Invalid key')
+        })
+      }
+
+      it('allow double dots in filename: ' , async () => {
+        await ctx.storage.setItem('s1/te..st..js', 'ok')
+        expect(await ctx.storage.getItem('s1/te..st..js')).toBe('ok')
       })
     }
   })