+++ test/drivers/memory.test.ts
@@ -1,3 +1,4 @@
+import { describe } from 'vitest'
 import driver from '../../src/drivers/memory'
 import { testDriver } from './utils'