+++ test/storage.test.ts
@@ -1,3 +1,4 @@
+import { describe, it, expect, vi } from 'vitest'
 import { createStorage, snapshot, restoreSnapshot, prefixStorage } from '../src'
 import memory from '../src/drivers/memory'
 
@@ -20,11 +21,67 @@ describe('storage', () => {
   })
 
   it('watch', async () => {
-    const onChange = jest.fn()
+    const onChange = vi.fn()
     const storage = createStorage().mount('/mnt', memory())
     await storage.watch(onChange)
     await restoreSnapshot(storage, data, 'mnt')
+    expect(onChange).toHaveBeenCalledWith('update', 'mnt:etc:conf')
     expect(onChange).toHaveBeenCalledWith('update', 'mnt:data:foo')
+    expect(onChange).toHaveBeenCalledTimes(2)
+  })
+
+  it('unwatch return', async () => {
+    const onChange = vi.fn()
+    const storage = createStorage().mount('/mnt', memory())
+    const unwatch = await storage.watch(onChange)
+    await storage.setItem('mnt:data:foo', 42)
+    await unwatch()
+    await storage.setItem('mnt:data:foo', 41)
+    expect(onChange).toHaveBeenCalledTimes(1)
+  })
+
+  it('unwatch all', async () => {
+    const onChange = vi.fn()
+    const storage = createStorage().mount('/mnt', memory())
+    await storage.watch(onChange)
+    await storage.setItem('mnt:data:foo', 42)
+    await storage.unwatch()
+    await storage.setItem('mnt:data:foo', 41)
+    expect(onChange).toHaveBeenCalledTimes(1)
+  })
+
+  it('mount overides', async () => {
+    const mainStorage = memory()
+    const storage = createStorage({ driver: mainStorage })
+    await storage.setItem('/mnt/test.txt', 'v1')
+    await storage.setItem('/mnt/test.base.txt', 'v1')
+
+    const initialKeys = await storage.getKeys()
+    expect(initialKeys).toMatchInlineSnapshot(`
+      [
+        "mnt:test.txt",
+        "mnt:test.base.txt",
+      ]
+    `)
+
+    storage.mount('/mnt', memory())
+    await storage.setItem('/mnt/test.txt', 'v2')
+
+    await storage.setItem('/mnt/foo/test.txt', 'v3')
+    storage.mount('/mnt/foo', memory())
+    expect(await storage.getItem('/mnt/foo/test.txt')).toBe(null)
+
+    expect(await storage.getItem('/mnt/test.txt')).toBe('v2')
+    expect(await storage.getKeys()).toMatchInlineSnapshot(`
+      [
+        "mnt:test.txt",
+      ]
+    `)
+
+    await storage.clear('/mnt')
+    await storage.unmount('/mnt')
+    expect(await storage.getKeys()).toMatchObject(initialKeys)
+    expect(await storage.getItem('/mnt/test.txt')).toBe('v1')
   })
 })